Move build-env-plugin.js
